remove old tests
