--- client_test/actors_test.go
@@ -1,10 +1,0 @@
-package client_test
-
-import "github.com/statechannels/go-nitro/internal/testactors"
-
-var (
-	alice = testactors.Alice
-	bob   = testactors.Bob
-	irene = testactors.Irene
-	brian = testactors.Brian
-)
