+++ client_test/crash_test.go
@@ -6,11 +6,14 @@ import (
 	"os"
 	"testing"
 
+	"github.com/ethereum/go-ethereum/common"
 	"github.com/statechannels/go-nitro/client"
 	"github.com/statechannels/go-nitro/client/engine"
 	"github.com/statechannels/go-nitro/client/engine/chainservice"
 	"github.com/statechannels/go-nitro/client/engine/messageservice"
 	"github.com/statechannels/go-nitro/client/engine/store"
+	ta "github.com/statechannels/go-nitro/internal/testactors"
+	"github.com/statechannels/go-nitro/internal/testdata"
 	"github.com/statechannels/go-nitro/rand"
 	"github.com/statechannels/go-nitro/types"
 	"github.com/tidwall/buntdb"
@@ -46,11 +49,11 @@ func TestCrashTolerance(t *testing.T) {
 	defer os.RemoveAll(dataFolder)
 
 	// Client setup
-	storeA := store.NewDurableStore(alice.PrivateKey, dataFolder, buntdb.Config{SyncPolicy: buntdb.Always})
-	messageserviceA := messageservice.NewTestMessageService(alice.Address(), broker, 0)
+	storeA := store.NewDurableStore(ta.Alice.PrivateKey, dataFolder, buntdb.Config{SyncPolicy: buntdb.Always})
+	messageserviceA := messageservice.NewTestMessageService(ta.Alice.Address(), broker, 0)
 	clientA := client.New(messageserviceA, chainA, storeA, logDestination, &engine.PermissivePolicy{}, nil)
 
-	clientB, _ := setupClient(bob.PrivateKey, chainB, broker, logDestination, 0)
+	clientB, _ := setupClient(ta.Bob.PrivateKey, chainB, broker, logDestination, 0)
 	defer closeClient(t, &clientB)
 	// End Client setup
 
@@ -59,9 +62,9 @@ func TestCrashTolerance(t *testing.T) {
 		channelId := directlyFundALedgerChannel(t, clientA, clientB, types.Address{})
 
 		closeClient(t, &clientA)
-		anotherMessageserviceA := messageservice.NewTestMessageService(alice.Address(), broker, 0)
+		anotherMessageserviceA := messageservice.NewTestMessageService(ta.Alice.Address(), broker, 0)
 		anotherChainA, err := chainservice.NewSimulatedBackendChainService(sim, bindings, ethAccounts[0], logDestination)
-		anotherStoreA := store.NewDurableStore(alice.PrivateKey, dataFolder, buntdb.Config{SyncPolicy: buntdb.Always})
+		anotherStoreA := store.NewDurableStore(ta.Alice.PrivateKey, dataFolder, buntdb.Config{SyncPolicy: buntdb.Always})
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -74,4 +77,24 @@ func TestCrashTolerance(t *testing.T) {
 		directlyDefundALedgerChannel(t, anotherClientA, clientB, channelId)
 
 	}
+}
+
+func directlyDefundALedgerChannel(t *testing.T, alpha client.Client, beta client.Client, channelId types.Destination) {
+	id := alpha.CloseLedgerChannel(channelId)
+	waitTimeForCompletedObjectiveIds(t, &alpha, defaultTimeout, id)
+	waitTimeForCompletedObjectiveIds(t, &beta, defaultTimeout, id)
+}
+
+const ledgerChannelDeposit = 5_000_000
+
+func directlyFundALedgerChannel(t *testing.T, alpha client.Client, beta client.Client, asset common.Address) types.Destination {
+	// Set up an outcome that requires both participants to deposit
+	outcome := testdata.Outcomes.Create(*alpha.Address, *beta.Address, ledgerChannelDeposit, ledgerChannelDeposit, asset)
+
+	response := alpha.CreateLedgerChannel(*beta.Address, 0, outcome)
+
+	<-alpha.ObjectiveCompleteChan(response.Id)
+	<-beta.ObjectiveCompleteChan(response.Id)
+
+	return response.ChannelId
 }
