+++ client_test/rpc_test.go
@@ -12,6 +12,7 @@ import (
 	p2pms "github.com/statechannels/go-nitro/client/engine/messageservice/p2p-message-service"
 	"github.com/statechannels/go-nitro/client/engine/store"
 	"github.com/statechannels/go-nitro/crypto"
+	ta "github.com/statechannels/go-nitro/internal/testactors"
 	"github.com/statechannels/go-nitro/internal/testdata"
 	"github.com/statechannels/go-nitro/rpc"
 	"github.com/statechannels/go-nitro/rpc/transport"
@@ -46,18 +47,18 @@ func executeRpcTest(t *testing.T, connectionType transport.TransportType) {
 	logDestination := newLogWriter(logFile)
 
 	chain := chainservice.NewMockChain()
-	chainServiceA := chainservice.NewMockChainService(chain, alice.Address())
-	chainServiceB := chainservice.NewMockChainService(chain, bob.Address())
-	chainServiceI := chainservice.NewMockChainService(chain, irene.Address())
+	chainServiceA := chainservice.NewMockChainService(chain, ta.Alice.Address())
+	chainServiceB := chainservice.NewMockChainService(chain, ta.Bob.Address())
+	chainServiceI := chainservice.NewMockChainService(chain, ta.Irene.Address())
 
-	rpcClientA, msgA, cleanupFnA := setupNitroNodeWithRPCClient(t, alice.PrivateKey, 3005, 4005, chainServiceA, logDestination, connectionType)
-	rpcClientB, msgB, cleanupFnB := setupNitroNodeWithRPCClient(t, bob.PrivateKey, 3006, 4006, chainServiceB, logDestination, connectionType)
-	rpcClientI, msgI, cleanupFnC := setupNitroNodeWithRPCClient(t, irene.PrivateKey, 3007, 4007, chainServiceI, logDestination, connectionType)
+	rpcClientA, msgA, cleanupFnA := setupNitroNodeWithRPCClient(t, ta.Alice.PrivateKey, 3005, 4005, chainServiceA, logDestination, connectionType)
+	rpcClientB, msgB, cleanupFnB := setupNitroNodeWithRPCClient(t, ta.Bob.PrivateKey, 3006, 4006, chainServiceB, logDestination, connectionType)
+	rpcClientI, msgI, cleanupFnC := setupNitroNodeWithRPCClient(t, ta.Irene.PrivateKey, 3007, 4007, chainServiceI, logDestination, connectionType)
 
 	peers := []p2pms.PeerInfo{
-		{Id: msgA.Id(), IpAddress: "127.0.0.1", Port: 3005, Address: alice.Address()},
-		{Id: msgB.Id(), IpAddress: "127.0.0.1", Port: 3006, Address: bob.Address()},
-		{Id: msgI.Id(), IpAddress: "127.0.0.1", Port: 3007, Address: irene.Address()},
+		{Id: msgA.Id(), IpAddress: "127.0.0.1", Port: 3005, Address: ta.Alice.Address()},
+		{Id: msgB.Id(), IpAddress: "127.0.0.1", Port: 3006, Address: ta.Bob.Address()},
+		{Id: msgI.Id(), IpAddress: "127.0.0.1", Port: 3007, Address: ta.Irene.Address()},
 	}
 	// Connect nitro P2P message services
 	msgA.AddPeers(peers)
@@ -68,8 +69,8 @@ func executeRpcTest(t *testing.T, connectionType transport.TransportType) {
 	defer cleanupFnB()
 	defer cleanupFnC()
 
-	res := rpcClientA.CreateLedger(irene.Address(), 100, testdata.Outcomes.Create(alice.Address(), irene.Address(), 100, 100, types.Address{}))
-	bobResponse := rpcClientB.CreateLedger(irene.Address(), 100, testdata.Outcomes.Create(bob.Address(), irene.Address(), 100, 100, types.Address{}))
+	res := rpcClientA.CreateLedger(ta.Irene.Address(), 100, testdata.Outcomes.Create(ta.Alice.Address(), ta.Irene.Address(), 100, 100, types.Address{}))
+	bobResponse := rpcClientB.CreateLedger(ta.Irene.Address(), 100, testdata.Outcomes.Create(ta.Bob.Address(), ta.Irene.Address(), 100, 100, types.Address{}))
 
 	// Quick sanity check that we're getting a valid objective id
 	assert.Regexp(t, "DirectFunding.0x.*", res.Id)
@@ -80,10 +81,10 @@ func executeRpcTest(t *testing.T, connectionType transport.TransportType) {
 	<-rpcClientI.ObjectiveCompleteChan(bobResponse.Id)
 
 	vRes := rpcClientA.CreateVirtual(
-		[]types.Address{irene.Address()},
-		bob.Address(),
+		[]types.Address{ta.Irene.Address()},
+		ta.Bob.Address(),
 		100,
-		testdata.Outcomes.Create(alice.Address(), bob.Address(), 100, 100, types.Address{}))
+		testdata.Outcomes.Create(ta.Alice.Address(), ta.Bob.Address(), 100, 100, types.Address{}))
 
 	assert.Regexp(t, "VirtualFund.0x.*", vRes.Id)
 
@@ -162,7 +163,7 @@ func setupNitroNodeWithRPCClient(
 	if err != nil {
 		panic(err)
 	}
-	rpcClient, err := rpc.NewRpcClient(rpcServer.Url(), alice.Address(), createLogger(logDestination, node.Address.Hex(), "client"), clienConnection)
+	rpcClient, err := rpc.NewRpcClient(rpcServer.Url(), ta.Alice.Address(), createLogger(logDestination, node.Address.Hex(), "client"), clienConnection)
 	if err != nil {
 		panic(err)
 	}
